# Render dump refactoring