--- a/chunky/src/java/se/llbit/chunky/renderer/RenderContext.java
+++ b/chunky/src/java/se/llbit/chunky/renderer/RenderContext.java
@@ -108,6 +108,7 @@ public class RenderContext {
    * @return A File object. Note, the file object may not exist yet and the directory leading to the file will be created.
    */
   public File getSceneFile(String fileName) {
+    ensureSceneDirectory();
     return new File(sceneDirectory, fileName);
   }
 
@@ -126,7 +127,6 @@ public class RenderContext {
    * @throws FileNotFoundException If the file does not exist.
    */
   public OutputStream getSceneFileOutputStream(String fileName) throws FileNotFoundException {
-    ensureSceneDirectory();
     return new FileOutputStream(getSceneFile(fileName));
   }