--- a/chunky/src/java/se/llbit/chunky/renderer/renderdump/CompressedFloatDumpFormat.java
+++ b/chunky/src/java/se/llbit/chunky/renderer/renderdump/CompressedFloatDumpFormat.java
@@ -1,4 +1,4 @@
-/* Copyright (c) 2012-2016 Jesper Öqvist <jesper@llbit.se>
+/* Copyright (c) 2021 Chunky contributors
  *
  * This file is part of Chunky.
  *