Fixes #783

This refactoring moves all the scene dump processing logic into it's own package.
The 3 primary use cases "load", "save", and "merge" can be found as static methods in the RenderDump class. It also handles the lookup of the correct DumpFormat strategy for the given dump format version.
The DumpFormat class provides different implementations for dump formats (including a registry for different dump format versions).