@aTom3333 2021-01-24
I feel that tracking progress at every pixel should have some noticeable overhead. Have you seen any difference in dump time when changing that? (or ideally done some measurements)
Ideally progress should be tracked for every row/column as it was done for the old format, that's just a bit tricky for the new one because it doesn't have a notion of row or column, it works on a linear stream.