@ShirleyNekoDev 2021-01-25
> I feel that tracking progress at every pixel should have some noticeable overhead. Have you seen any difference in dump time when changing that? (or ideally done some measurements)

I did a quick check and noticed a big slow down (more than 10x worse).
Theoretically this shouldn't have that much of an impact for JavaFX, but I didn't check the indirection in the TaskTracker and it's listeners.

> Ideally progress should be tracked for every row/column as it was done for the old format, that's just a bit tricky for the new one because it doesn't have a notion of row or column, it works on a linear stream.

A fix for this will come with my next commit by reporting progress (you named it) only once per row again.