@ShirleyNekoDev 2021-01-25
Performance is now back for large dumps - thanks for the warning!

Scene size: 2000x2000

## Original code

- Dump format version 1
```
Loading took 266ms
Saving took 260ms

Loading took 265ms
Saving took 242ms
```

## New code

- Dump format version 1
```
Loading took 273ms
Saving took 260ms

Loading took 274ms
Saving took 261ms

Loading took 287ms
Saving took 276ms
```

- Dump format version 0
```
Loading took 12032ms
Saving took 17039ms

Loading took 12213ms
Saving took 16923ms

Loading took 12111ms
Saving took 16887ms
```

- Dump format version 0 (Without sending progress via pixelProgress)
```
Loading took 12010ms
Saving took 17020ms
```